small refactor ; restore .gitignore ; remove dependency on ConstraintLayout (for jenkins)
