+++ app/src/main/res/layout/activity_displaychain.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<android.widget.LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -18,4 +18,4 @@
             android:layout_height="wrap_content"
             android:text="loading..." />
     </ScrollView>
-</android.support.constraint.ConstraintLayout>
+</android.widget.LinearLayout>
